Massively rework the Docker images
They should be much more cache efficient now, perform correctly (always build
the code by COPYing it in instead of using git clone), and are based on
upstream ROS images.

There are additionally two images: a dev one that requires code to be compiled
in the image before use and a prod one that has everything compiled already.

See the README for more details.