deadreckoning: fix crash on surface and actually update dr_msg when underwater
`NavSatFix.status.STATUS_SBAS_FIX` was an incorrect way to access
`STATUS_SBAS_FIX` and it caused the deadreckoning node to crash when the glider
hit the surface. Fix by using `NavSatStatus.STATUS_SBAS_FIX` instead.

Second, the code to update the dr position based on GPS output was always
triggered on every loop as soon as the first GPS fix was received. Fix by
clearing the gps_msg (setting it to None) after we're done with it. Chose this
method instead of the half implemented new_gps flag because now if gps_msg is
ever accessed when it shouldn't be, we should get a nice crash.