Fix bugs with GPS update rate
The update rate of the GPS is specified in the xacro file using the updateRate
tag. This code was looking for update_rate.

Additionally, the update_rate_ var was never initialized, meaning that its
intial value (and only value due to the incorrect tag name) was whatever junk
was present on the heap when the class was instantiated. On my computer, when
running in Docker this was consistenly a value with a random mantissa but
exponent always around 10^-310. That is an absurdly slow update rate. However,
another bug is that updateTimer.setUpdateRate was called with a period instead
of rate. So this absurdly slow update rate got turned into an absurdly fast
update rate when configuring the timer.

When running outside of Docker, for whatever reason, I consistently got
update_rate_ initialized to zero. This turned signal_confirm_time_ into a NaN,
resulting in a GPS update never being sent.

While I was at it, I also initialized some more of the class variables.