free temporary coordinate transformations when we're done with them
Discovered with valgrind

Prior to this change, Gazebo leaked about 1 GB of RAM for every hour of
simulation time when sitting on the surface. After this change, I ran Gazebo
for 24 hours of simulation time and saw stable memory usage from the gzserver
process.

The next obvious source of memory troubles is going to be the map viewer. I
didn't see any easy way to clear out old popups in batch.